Accommodate strings that get split by onCharacters
When there is &quot; in the string the xml parser splits the string into different parts. This makes the opensrs package bug too when returning "response_text" property. For example when the "response_text" is 'DNS Zone could not be set for qpwoeip18301.com. (Record type "" is not valid or supported.)' it returns only ' is not valid or supported.)' which is a problem for me since I'd like to show errors from the OpenSRS API directly to clients. (Too lazy to think of proper errors messages myself). On the other hand I was not too lazy to fix the code (which only took like 2 hours to figure out.. so here you go. Hope it gets into the repo. Or if you know of a cleverer way to do it - please tell me so I fix my code.